# projeto_sistema_pedidos.swagger.yaml
info:
  title: projeto_sistema_pedidos
  description: ''
  version: 1.0.0
...
paths:
...
  /api/produtos:
    get:
      summary: Listar todos os produtos
      deprecated: false
      description: ''
      operationId: get_app_produto_listar
      tags:
        - Produtos
        - Produtos
      parameters: []
      responses:
        '200':
          description: Lista de produtos
          headers: {}
          schema:
            type: array
            items:
              properties:
                id:
                  type: integer
                nome:
                  type: string
                descricao:
                  type: string
                preco:
                  type: number
                  format: float
                categoria:
                  type: string
              type: object
      security: []
      produces:
        - application/json
    post:
      summary: Criar um novo produto
      deprecated: false
      description: ''
      operationId: post_app_produto_criar
      tags:
        - Produtos
        - Produtos
      parameters:
        - name: body
          in: body
          schema:
            properties:
              nome:
                type: string
              descricao:
                type: string
              preco:
                type: number
                format: float
              categoria:
                type: string
            type: object
      responses:
        '201':
          description: Produto criado com sucesso
          headers: {}
          schema:
            properties:
              id:
                type: integer
              nome:
                type: string
              descricao:
                type: string
              preco:
                type: number
                format: float
              categoria:
                type: string
            type: object
      security: []
      consumes:
        - application/json
      produces:
        - application/json
  /api/produtos/{id}:
    put:
      summary: Atualizar um produto
      deprecated: false
      description: ''
      operationId: put_app_produto_atualizar
      tags:
        - Produtos
        - Produtos
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          type: integer
          x-example: 0
        - name: body
          in: body
          schema:
            properties:
              nome:
                type: string
              descricao:
                type: string
              preco:
                type: number
                format: float
              categoria:
                type: number
            type: object
      responses:
        '200':
          description: Produto atualizado com sucesso
          headers: {}
          schema:
            properties:
              id:
                type: integer
              nome:
                type: string
              descricao:
                type: string
              preco:
                type: number
                format: float
              categoria:
                type: string
            type: object
      security: []
      consumes:
        - application/json
      produces:
        - application/json
    delete:
      summary: Deletar um produto
      deprecated: false
      description: ''
      operationId: delete_app_produto_deletar
      tags:
        - Produtos
        - Produtos
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          type: integer
          x-example: 0
      responses:
        '200':
          description: Produto deletado com sucesso
          headers: {}
          schema:
            properties:
              message:
                type: string
            type: object
        '404':
          description: Produto não encontrado
          headers: {}
          schema:
            properties:
              error:
                type: string
            type: object
      security: []
      produces:
        - application/json
  /api/produtos/{categoria}:
    get:
      summary: Listar produtos por categoria
      deprecated: false
      description: ''
      operationId: get_app_produto_listarporcategoria
      tags:
        - Produtos
        - Produtos
      parameters:
        - name: categoria
          in: path
          description: ''
          required: true
          type: integer
          x-example: 0
      responses:
        '200':
          description: Lista de produtos por categoria
          headers: {}
          schema:
            type: array
            items:
              properties:
                id:
                  type: integer
                nome:
                  type: string
                descricao:
                  type: string
                preco:
                  type: number
                  format: float
                categoria:
                  type: string
              type: object
      security: []
      produces:
        - application/json
...
swagger: '2.0'
...
securityDefinitions: {}
...